pyproject.toml
[tool.poetry]
name = "MapAnalyzer"
version = "0.0.87"
description = ""
authors = ["Your Name <you@example.com>"]
license = "GNU"
readme = "README.md"

[tool.poetry.build]
script = "build.py"
generate-setup-file = true

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
burnysc2 = "^6.4.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
hypothesis = "^6.82.7"
pytest-cov = "^4.1.0"
matplotlib = "^3.7.2"
click = "^8.1.7"
pytest-html = "^3.2.0"
monkeytype = "^23.3.0"
pytest-benchmark = "^4.0.0"
coverage = "^7.3.0"
codecov = "^2.1.13"
scikit-image = "^0.21.0"
isort = "^5.12.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core", "numpy", "setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

